requirements.txt
langchain
langchain-google-genai
langgraph
pydantic
datasets
pandas
rouge-score
streamlit
python-dotenv